refactor(agent): merge conversation-supervisor into conversation domain
