+++ packages/agent/src/conversation/index.ts
@@ -156,3 +156,16 @@ export namespace ConversationHandler {
     };
   }
 }
+
+// Re-export conversation-supervisor
+export {
+  ConversationSupervisor,
+  type ConversationSupervisorConfig,
+  type ConversationInput,
+  type ConversationPlan,
+  type WorkItemOutline,
+  type ApprovalDecision,
+  type ExecutionContextFork,
+  type ConversationHistory,
+  type ConversationSupervisorResult,
+} from "./conversation-supervisor";
