+++ packages/agent/src/ingress/run-executor.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from "crypto";
 import { TaskManager } from "../task";
-import { ConversationSupervisor } from "../loop";
+import { ConversationSupervisor } from "../conversation";
 import { ExecutionSupervisor } from "../execution";
 import { classifyLane } from "./event-kinds";
 import type { RunRequest, RunResult } from "./engine";
