+++ packages/agent/src/loop/index.ts
@@ -1,13 +1,1 @@
-// Loop module — conversation supervisor
-
-export {
-  ConversationSupervisor,
-  type ConversationSupervisorConfig,
-  type ConversationInput,
-  type ConversationPlan,
-  type WorkItemOutline,
-  type ApprovalDecision,
-  type ExecutionContextFork,
-  type ConversationHistory,
-  type ConversationSupervisorResult,
-} from "./conversation-supervisor";
+// Loop module — execution supervisors and workers
